Move blockstore purge methods to submodule
